Fix compiler warning.

+++ examples/utils/snitch.cpp
@@ -381,9 +381,9 @@ void compress(State& state, const std::string& folder, const std::string& archiv
         block.uncompressed = file.size();
 
         // write the file in multiple parts
-        for (s64 offset = 0; offset < file.size(); offset += large_block_size)
+        for (u64 offset = 0; offset < file.size(); offset += large_block_size)
         {
-            u64 size = std::min(large_block_size, u64(file.size() - offset));
+            u64 size = std::min(large_block_size, file.size() - offset);
 #ifdef DISABLE_MMAP
             file.read(buffer.data(), size);
             output.write(buffer.data(), size);
